requirements.txt
python-dotenv
groq
pygame
edge-tts
pyttsx3
cohere
flask
rich